Reinformence Learning based on Robocode
Under LUTRobot folder:
An implementation of Reinforcement Learning based on Robocode environment.  In this implementation, we train the robot based on reinforcement learning and save the data to LUT table.

Under measure preformande with rounds 4000:
In this implementation, the code is automatically set to verify the performance of trained robot under Sarsa and termimal reward only situations.